PR: Added tools to support migration
Updated the migrate.py to be weirdly python3 compliant and to
work with environment variables if present. Hashlib had to be imported
otherwise the connector would Seg fault (cool work package mysql
package maintainer dudes!)

Added a check at the end for automated runs so it doesn't prompt you to close the window...
though, to be honest, this input step should really be removed regardless (must be a windows things)
and if that's the case, add a proper check for the OS that it is running in.

Included a Dockerfile and docker-compose to support migrations
in a Jenkins pipeline by installing the necessary python tools
and mounted files to run the migration script.